refactor (delivery) : delivery manager 추가를 위한 delivery 구조 변경 (#14)
