+++ src/main/java/com/logilink/eureka/client/delivery/domain/delivery/model/dto/responseDto/DeleteResponseDto.java
@@ -1,15 +1,16 @@
-package com.logilink.eureka.client.delivery.domain.model.dto.responseDto;
+package com.logilink.eureka.client.delivery.domain.delivery.model.dto.responseDto;
 
 import com.logilink.eureka.client.delivery.common.constants.DeliveryStatus;
-import com.logilink.eureka.client.delivery.domain.model.entity.Delivery;
+import com.logilink.eureka.client.delivery.domain.delivery.model.entity.Delivery;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Getter
 @Setter
-public class ResponseDto {
+public class DeleteResponseDto {
     private UUID deliveryId;
 
     private boolean isHubDelivery;
@@ -28,7 +29,11 @@ public class ResponseDto {
 
     private UUID orderId;
 
-    public ResponseDto(Delivery delivery) {
+    private LocalDateTime deletedAt;
+
+    private Long deletedBy;
+
+    public DeleteResponseDto(Delivery delivery) {
         this.deliveryId = delivery.getId();
         this.isHubDelivery = delivery.isHubDelivery();
         this.status = delivery.getStatus();
@@ -38,5 +43,8 @@ public class ResponseDto {
         this.deliveryManagerId = delivery.getDeliveryManagerId();
         this.routeId = delivery.getRouteId();
         this.orderId = delivery.getOrderId();
+        this.deletedAt = delivery.getDeletedAt();
+        this.deletedBy = delivery.getDeletedBy();
     }
+
 }
