+++ src/main/java/com/logilink/eureka/client/delivery/domain/delivery/model/dto/responseDto/SearchDeliveryResponseDto.java
@@ -1,4 +1,4 @@
-package com.logilink.eureka.client.delivery.domain.model.dto.responseDto;
+package com.logilink.eureka.client.delivery.domain.delivery.model.dto.responseDto;
 
 import com.logilink.eureka.client.delivery.common.constants.DeliveryStatus;
 import lombok.AllArgsConstructor;
